+++ src/FlatfileExportConfiguration.php
@@ -30,6 +30,8 @@ class FlatfileExportConfiguration
 
         // Normalize the different ways to make field specifications
         $this->fields = collect($fields)->map(function ($value, $key) {
+            $callback = null;
+
             if (is_array($value)) {
                 if (! Arr::exists($value, 'column')) {
                     return Arr::add($value, 'column', $key);
@@ -38,11 +40,28 @@ class FlatfileExportConfiguration
                 return $value;
             }
 
+            if (is_callable($value)) {
+                $callback = $value;
+                $value = $key;
+            }
+
+            if (is_numeric($key)) {
+                $key = $value;
+            }
+
+            if ($callback) {
+                return [
+                    'column'   => $key,
+                    'label'    => $value,
+                    'callback' => $callback,
+                ];
+            }
+
             return [
-                'column'   => $key,
-                'label'    => $value,
+                'column' => $key,
+                'label'  => $value,
             ];
-        });
+        })->values();
 
         return $this;
     }