+++ changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Support to add Collection to addRow() to be used in custom queries
+
 ### Changed
 - Callbacks now also get the complete "row"-model to process cross-column informations: `function ($value, $model) {}`